Move hold-to-confirm back to UI section
